kr3/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
passlib>=1.7.4
bcrypt>=4.0.0
pyjwt>=2.8.0
python-dotenv>=1.0.0
slowapi>=0.1.9
itsdangerous>=2.1.0
email-validator>=2.0.0
